backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
